Handle scenario where there is a user without profile

+++ users/controller.js
@@ -1,22 +1,27 @@
 const UserModel = require('./models/UserModel');
-const { NoUserFound } = require('./errors');
+const { NoUserFound, NoProfileFound } = require('./errors');
 const nativeEager = require('../util/nativeEager');
 
 class UsersController {
   static async delete(id, cascade, builder = UserModel.query()) {
+    // TODO : replace this findById by this.load so we can generalize the profile error loading
     const userProfile = await builder
       .eager('profile')
       .findById(id);
+    if (!userProfile) throw NoUserFound;
+    if (!userProfile.profile) throw NoProfileFound;
     if (userProfile.profile.hasChildren() && cascade) {
       await Promise.all(userProfile.profile.children.map(
         childUserId => UsersController.delete(childUserId, false)));
     }
+    // TODO : delete reference from the parent as well
     await userProfile.profile.$query().delete();
     return userProfile.$query().delete();
   }
 
   // does not check if exists, please do beforeHand
   static async softDelete(id, cascade, builder = UserModel.query()) {
+    // TODO : replace this qb.where by this.load so we can generalize the profile error loading
     const res = await builder
       .allowEager('[profile]')
       .eager('profile')
@@ -24,6 +29,8 @@ class UsersController {
       .softDelete()
       .returning('*');
     const userProfile = res[0];
+    if (!userProfile) throw NoUserFound;
+    if (!userProfile.profile) throw NoProfileFound;
     userProfile.profile = await userProfile.profile.$query().softDelete().returning('*');
     if (userProfile.profile.hasChildren() && cascade) {
       await Promise.all(userProfile.profile.children.map(
